fix: enhance hostname sanitization and refactor OSM API collections
