+++ src/utils/whitelabel.ts
@@ -11,15 +11,24 @@ export async function getWhitelabelConfig(hostname: string) {
 }
 
 export function sanitizeHostname(hostnameOrIPAddress: string) {
-  return (
-    hostnameOrIPAddress
-      // Allow sharing a deployment with ngrok
-      .replace(/.*\.ngrok(?:-free)\.app$/, "wheelmap.tech")
-      // Allow branch test deployments with a common branding
-      .replace(/.*\.wheelmap\.tech$/, "wheelmap.tech")
+  let hostname = hostnameOrIPAddress
+    // Allow sharing a deployment with ngrok
+    .replace(/.*\.ngrok(?:-free)\.app$/, "wheelmap.tech")
+    // Allow branch test deployments with a common branding
+    .replace(/.*\.wheelmap\.tech$/, "wheelmap.tech");
+
+  // Only replace local/private network IPs in development mode
+  if (process.env.NODE_ENV === "development") {
+    hostname = hostname
       // Use 'localhost' branding for loopback IPs
       .replace(/127\.0\.0\.1/, "localhost")
       .replace(/0\.0\.0\.0/, "localhost")
       .replace(/::1/, "localhost")
-  );
+      // Use 'localhost' branding for private network IPs (192.168.x.x, 10.x.x.x, 172.16-31.x.x)
+      .replace(/^192\.168\.\d{1,3}\.\d{1,3}$/, "localhost")
+      .replace(/^10\.\d{1,3}\.\d{1,3}\.\d{1,3}$/, "localhost")
+      .replace(/^172\.(1[6-9]|2\d|3[01])\.\d{1,3}\.\d{1,3}$/, "localhost");
+  }
+
+  return hostname;
 }
